# src/docs/openapi.yaml
openapi: 3.0.0
info:
    title: Multi-Client REST API
    version: 1.0.0
servers:
    - url: http://localhost:3000/api
paths:
...
    /games_play/admin/games:
        get:
            summary: Връща всички игри с пейджинг (само за администратори)
            tags: [GamesPlay Admin]
            security:
                - gamesPlay: []
            parameters:
                - in: query
                  name: page
                  schema:
                      type: integer
                      default: 1
                      minimum: 1
                  description: Номер на страницата (лимитът е фиксиран на 5)
            responses:
                200:
                    description: Успешно взети игри
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GamesPlayAdminGamesListResponse"
                401:
                    description: Unauthorized - Липсващ токен
                    content:
                        application/json:
                            example:
                                message: "Invalid token!"
                403:
                    description: Forbidden - изисква се администраторски достъп
                    content:
                        application/json:
                            example:
                                message: "Admin access required"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
    /games_play/admin/games/{gameId}:
        delete:
            summary: Изтрива игра по ID (само за администратори)
            tags: [GamesPlay Admin]
            security:
                - gamesPlay: []
            parameters:
                - in: path
                  name: gameId
                  required: true
                  schema:
                      type: string
                  description: ID на играта (валиден ObjectId – 24 символа)
            responses:
                204:
                    description: No Content - Успешно изтрита игра
                400:
                    description: Bad Request - Невалиден ID
                    content:
                        application/json:
                            example:
                                message: "Id must be a valid MongooseDB ObjectId!"
                401:
                    description: Unauthorized - Липсващ токен
                    content:
                        application/json:
                            example:
                                message: "Invalid token!"
                403:
                    description: Forbidden - изисква се администраторски достъп
                    content:
                        application/json:
                            example:
                                message: "Admin access required"
                404:
                    description: Not Found - Играта не съществува
                    content:
                        application/json:
                            example:
                                message: "Game not found!"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
...
    /games_play/games:
        get:
            summary: Връща всички игри с опция лимит на респонса и филтриране по заглавие
            tags: [GamesPlay Games]
            parameters:
                - in: query
                  name: search
                  schema:
                      type: string
                  required: false
                  description: Търсене по заглавие на играта

                - in: query
                  name: limit
                  schema:
                      type: integer
                  required: false
                  description: Максимален брой игри в отговора
            responses:
                200:
                    description: Успешно взети игри
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GamesPlayGamesGamesListResponse"
                400:
                    description: Bad Request - Невалиден лимит
                    content:
                        application/json:
                            example:
                                message: "Limit must be a positive integer!"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
        post:
            summary: Създава нова игра
            tags: [GamesPlay Games]
            security:
                - gamesPlay: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/GamesPlayGamesCreateGameRequest"
            responses:
                201:
                    description: Успешно създадена игра
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GamesPlayGamesCreateGameResponse"
                400:
                    description: Bad Request - Невалидни данни
                    content:
                        application/json:
                            examples:
                                invalidTitle:
                                    summary: Invalid title
                                    value:
                                        message: "Game title should be at least 3 characters long!"
                                invalidCategory:
                                    summary: Invalid category
                                    value:
                                        message: "Category should be at least 3 characters long!"
                                invalidMaxLevel:
                                    summary: Invalid max level
                                    value:
                                        message: "MaxLevel...."
                                invalidImgUrl:
                                    summary: Invalid image URL
                                    value:
                                        message: "Invalid image url!"
                                invalidSummary:
                                    summary: Invalid summary
                                    value:
                                        message: "Summary should be at least 10 characters long!"
                401:
                    description: Unauthorized - Липсващ токен
                    content:
                        application/json:
                            example:
                                message: "Invalid token!"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
    /games_play/games/infinity:
        get:
            summary: Връща всички игри с пейджинг
            tags: [GamesPlay Games]
            parameters:
                - in: query
                  name: page
                  schema:
                      type: integer
                      default: 1
                      minimum: 1
                  description: Номер на страницата (лимитът е фиксиран на 5)
            responses:
                200:
                    description: Успешно взети игри
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GamesPlayAdminGamesListResponse"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
    /games_play/games/last_three:
        get:
            summary: Връща последните три игри
            tags: [GamesPlay Games]
            responses:
                200:
                    description: Успешно взети игри
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GamesPlayGamesGamesLast_ThreeResponse"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
    /games_play/games/{gameId}:
        get:
            summary: Взема игра по ID
            tags: [GamesPlay Games]
            parameters:
                - in: path
                  name: gameId
                  required: true
                  schema:
                      type: string
                  description: ID на играта (валиден ObjectId – 24 символа)
            responses:
                200:
                    description: Успешно вземане на коментари
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GamesPlayGamesCreateGameResponse"
                400:
                    description: Bad Request - Невалиден ID
                    content:
                        application/json:
                            example:
                                message: "Id must be a valid MongooseDB ObjectId!"
                404:
                    description: Not Found - Играта не съществува
                    content:
                        application/json:
                            example:
                                message: "There is no game with this id!"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
        put:
            summary: Редактира игра по ID (само за автори)
            tags: [GamesPlay Games]
            security:
                - gamesPlay: []
            parameters:
                - in: path
                  name: gameId
                  required: true
                  schema:
                      type: string
                  description: ID на играта (валиден ObjectId – 24 символа)
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/GamesPlayGamesCreateGameRequest"
            responses:
                201:
                    description: Успешно редактирана игра
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GamesPlayGamesCreateGameResponse"
                400:
                    description: Bad Request - Невалидни данни
                    content:
                        application/json:
                            examples:
                                invalidGameId:
                                    summary: Invalid game ID
                                    value:
                                        message: "Id must be a valid MongooseDB ObjectId!"
                                invalidTitle:
                                    summary: Invalid title
                                    value:
                                        message: "Game title should be at least 3 characters long!"
                                invalidCategory:
                                    summary: Invalid category
                                    value:
                                        message: "Category should be at least 3 characters long!"
                                invalidMaxLevel:
                                    summary: Invalid max level
                                    value:
                                        message: "MaxLevel...."
                                invalidImgUrl:
                                    summary: Invalid image URL
                                    value:
                                        message: "Invalid image url!"
                                invalidSummary:
                                    summary: Invalid summary
                                    value:
                                        message: "Summary should be at least 10 characters long!"
                401:
                    description: Unauthorized - Липсващ токен
                    content:
                        application/json:
                            example:
                                message: "Invalid token!"
                403:
                    description: Forbidden - само авторът на играта може да я редактира
                    content:
                        application/json:
                            example:
                                message: "You are not the owner of this resource!"
                404:
                    description: Not Found - Играта не съществува
                    content:
                        application/json:
                            examples:
                                invalidGameIdMiddleware:
                                    summary: Invalid resource ID
                                    value:
                                        message: "Resource not found!"
                                invalidGameUpdate:
                                    summary: Invalid update
                                    value:
                                        message: "Game not found!"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
        delete:
            summary: Изтрива игра по ID (само за автори)
            tags: [GamesPlay Games]
            security:
                - gamesPlay: []
            parameters:
                - in: path
                  name: gameId
                  required: true
                  schema:
                      type: string
                  description: ID на играта (валиден ObjectId – 24 символа)
            responses:
                204:
                    description: No Content - Успешно изтрита игра
                400:
                    description: Bad Request - Невалидни данни
                    content:
                        application/json:
                            example:
                                message: "Id must be a valid MongooseDB ObjectId!"
                401:
                    description: Unauthorized - Липсващ токен
                    content:
                        application/json:
                            example:
                                message: "Invalid token!"
                403:
                    description: Forbidden - само авторът на играта може да я редактира
                    content:
                        application/json:
                            example:
                                message: "You are not the owner of this resource!"
                404:
                    description: Not Found - Играта не съществува
                    content:
                        application/json:
                            examples:
                                invalidGameIdMiddleware:
                                    summary: Invalid resource ID
                                    value:
                                        message: "Resource not found!"
                                invalidGameUpdate:
                                    summary: Invalid update
                                    value:
                                        message: "Game not found!"
                500:
                    description: Internal Server Error - Вътрешна грешка на сървъра
                    content:
                        application/json:
                            example:
                                message: "Internal server error..."
...
components:
    securitySchemes:
        gamesPlay:
            type: apiKey
            in: cookie
            name: auth_GamesPlay
        cookingTogether:
            type: apiKey
            in: cookie
            name: auth_cooking
        classBook:
            type: apiKey
            in: cookie
            name: auth
    schemas:
...
        GamesPlayAdminGamesListResponse:
            type: object
            properties:
                games:
                    type: array
                    items:
                        $ref: "#/components/schemas/GamesPlayGame"
        GamesPlayGame:
            type: object
            properties:
                _id:
                    type: string
                    example: "6833318fd414ea08fac843ec"
                title:
                    type: string
                    example: "Test"
                category:
                    type: string
                    example: "Test category"
                maxLevel:
                    type: integer
                    example: 95
                imageUrl:
                    type: string
                    format: uri
                    example: "https://example.com/image.jpg"
                summary:
                    type: string
                _ownerId:
                    $ref: "#/components/schemas/GamesPlayUserBasic"
                createdAt:
                    type: string
                    format: date-time
                    example: "2025-05-25T15:04:47.780Z"
                updatedAt:
                    type: string
                    format: date-time
                    example: "2025-05-25T15:04:47.784Z"
        GamesPlayUserBasic:
            type: object
            properties:
                _id:
                    type: string
                    example: "6833307c97d956479e7c161e"
                email:
                    type: string
                    example: "user@example.com"
                role:
                    type: string
                    example: "user"
                dateCreated:
                    type: string
                    format: date-time
                    example: "2025-05-25T15:00:12.341Z"
...
        GamesPlayGamesGamesListResponse:
            type: array
            items:
                $ref: "#/components/schemas/GamesPlayGame"
        GamesPlayGamesCreateGameRequest:
            type: object
            required:
                - title
                - category
                - maxLevel
                - imageUrl
                - summary
            properties:
                title:
                    type: string
                    example: "New Game Title"
                category:
                    type: string
                    example: "Action"
                maxLevel:
                    type: integer
                    example: 50
                imageUrl:
                    type: string
                    format: uri
                    example: "https://example.com/game-image.jpg"
                summary:
                    type: string
                    example: "This is a brief summary of the game."
        GamesPlayGamesCreateGameResponse:
            type: object
            properties:
                _id:
                    type: string
                    example: "60c72b2f9b1e8c001f8e4d3a"
                title:
                    type: string
                    example: "New Game Title"
                category:
                    type: string
                    example: "Action"
                maxLevel:
                    type: integer
                    example: 50
                imageUrl:
                    type: string
                    format: uri
                    example: "https://example.com/game-image.jpg"
                summary:
                    type: string
                    example: "This is a brief summary of the game."
                _ownerId:
                    type: string
                    example: "60c72b2f9b1e8c001f8e4d3a"
                createdAt:
                    type: string
                    format: date-time
                    example: "2021-06-14T12:00:00Z"
                updatedAt:
                    type: string
                    format: date-time
                    example: "2021-06-14T12:00:00Z"
        GamesPlayGamesGamesLast_ThreeResponse:
            type: array
            items:
                $ref: "#/components/schemas/GamesPlayGamesCreateGameResponse"